Adds server attributes for elevated users
Adds the start of elevated attributes on servers. Instead of
having a separate admin API or muddying the API with extensions,
we add a simple "_elevated" attribute to the server resource that
we can hang additional attributes only exposed to users with
elevated permissions.

Also adds a bunch of inline server attribute documentation that
will be useful when translating all of it to JSON-Schema.